+++ wp-content/plugins/vf-gutenberg/assets/vf-blocks.js
@@ -1796,6 +1796,18 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var core
 
 /***/ }),
 
+/***/ "./wp-content/plugins/vf-gutenberg/blocks/vf-core/transforms/grid.js":
+/*!***************************************************************************!*\
+  !*** ./wp-content/plugins/vf-gutenberg/blocks/vf-core/transforms/grid.js ***!
+  \***************************************************************************/
+/*! exports provided: fromColumns */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"fromColumns\", function() { return fromColumns; });\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"react\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var _wordpress_blocks__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! @wordpress/blocks */ \"@wordpress/blocks\");\n/* harmony import */ var _wordpress_blocks__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(_wordpress_blocks__WEBPACK_IMPORTED_MODULE_1__);\n/**\n * Block transforms for: `vf/grid`, `vf/embl-grid`, and `core/columns`\n */\n\n\n\n// Map inner blocks and create new `vf/grid-column` with their children\n// New columns are appended to match minimum\n// End columns are merged to match maximum\nconst fromColumns = (fromBlock, toBlock, min, max) => {\n  return {\n    type: 'block',\n    blocks: [fromBlock],\n    // Match function (ignore initial placeholder state)\n    isMatch: attributes => attributes.placeholder !== 1,\n    // Transform function\n    transform: (attributes, innerBlocks) => {\n      // Map column props\n      let innerProps = innerBlocks.map(block => ({\n        attributes: {...block.attributes},\n        innerBlocks: [...block.innerBlocks]\n      }));\n      // Fill empty props to match min number of columns\n      while (innerProps.length < min) {\n        innerProps.push({});\n      }\n      // Merge end props to match max number of columns\n      while (innerProps.length > max) {\n        const mergeProps = innerProps.pop();\n        innerProps[innerProps.length - 1].innerBlocks.push(\n          ...mergeProps.innerBlocks\n        );\n      }\n      // Return new grid block with inner columns\n      return Object(_wordpress_blocks__WEBPACK_IMPORTED_MODULE_1__[\"createBlock\"])(\n        toBlock,\n        {columns: innerProps.length},\n        innerProps.map(props =>\n          Object(_wordpress_blocks__WEBPACK_IMPORTED_MODULE_1__[\"createBlock\"])(\n            'vf/grid-column',\n            props.attributes || {},\n            props.innerBlocks || []\n          )\n        )\n      );\n    }\n  };\n};\n\n\n//# sourceURL=webpack:///./wp-content/plugins/vf-gutenberg/blocks/vf-core/transforms/grid.js?");
+
+/***/ }),
+
 /***/ "./wp-content/plugins/vf-gutenberg/blocks/vf-core/vf-activity-item.jsx":
 /*!*****************************************************************************!*\
   !*** ./wp-content/plugins/vf-gutenberg/blocks/vf-core/vf-activity-item.jsx ***!
@@ -1912,7 +1924,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var reac
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var core_js_modules_es_symbol__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! core-js/modules/es.symbol */ \"./node_modules/core-js/modules/es.symbol.js\");\n/* harmony import */ var core_js_modules_es_symbol__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_symbol__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var core_js_modules_es_symbol_description__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! core-js/modules/es.symbol.description */ \"./node_modules/core-js/modules/es.symbol.description.js\");\n/* harmony import */ var core_js_modules_es_symbol_description__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_symbol_description__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var core_js_modules_es_symbol_iterator__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! core-js/modules/es.symbol.iterator */ \"./node_modules/core-js/modules/es.symbol.iterator.js\");\n/* harmony import */ var core_js_modules_es_symbol_iterator__WEBPACK_IMPORTED_MODULE_2___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_symbol_iterator__WEBPACK_IMPORTED_MODULE_2__);\n/* harmony import */ var core_js_modules_es_array_fill__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! core-js/modules/es.array.fill */ \"./node_modules/core-js/modules/es.array.fill.js\");\n/* harmony import */ var core_js_modules_es_array_fill__WEBPACK_IMPORTED_MODULE_3___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_array_fill__WEBPACK_IMPORTED_MODULE_3__);\n/* harmony import */ var core_js_modules_es_array_filter__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! core-js/modules/es.array.filter */ \"./node_modules/core-js/modules/es.array.filter.js\");\n/* harmony import */ var core_js_modules_es_array_filter__WEBPACK_IMPORTED_MODULE_4___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_array_filter__WEBPACK_IMPORTED_MODULE_4__);\n/* harmony import */ var core_js_modules_es_array_from__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! core-js/modules/es.array.from */ \"./node_modules/core-js/modules/es.array.from.js\");\n/* harmony import */ var core_js_modules_es_array_from__WEBPACK_IMPORTED_MODULE_5___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_array_from__WEBPACK_IMPORTED_MODULE_5__);\n/* harmony import */ var core_js_modules_es_array_iterator__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! core-js/modules/es.array.iterator */ \"./node_modules/core-js/modules/es.array.iterator.js\");\n/* harmony import */ var core_js_modules_es_array_iterator__WEBPACK_IMPORTED_MODULE_6___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_array_iterator__WEBPACK_IMPORTED_MODULE_6__);\n/* harmony import */ var core_js_modules_es_object_assign__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! core-js/modules/es.object.assign */ \"./node_modules/core-js/modules/es.object.assign.js\");\n/* harmony import */ var core_js_modules_es_object_assign__WEBPACK_IMPORTED_MODULE_7___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_object_assign__WEBPACK_IMPORTED_MODULE_7__);\n/* harmony import */ var core_js_modules_es_object_get_own_property_descriptor__WEBPACK_IMPORTED_MODULE_8__ = __webpack_require__(/*! core-js/modules/es.object.get-own-property-descriptor */ \"./node_modules/core-js/modules/es.object.get-own-property-descriptor.js\");\n/* harmony import */ var core_js_modules_es_object_get_own_property_descriptor__WEBPACK_IMPORTED_MODULE_8___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_object_get_own_property_descriptor__WEBPACK_IMPORTED_MODULE_8__);\n/* harmony import */ var core_js_modules_es_object_get_own_property_descriptors__WEBPACK_IMPORTED_MODULE_9__ = __webpack_require__(/*! core-js/modules/es.object.get-own-property-descriptors */ \"./node_modules/core-js/modules/es.object.get-own-property-descriptors.js\");\n/* harmony import */ var core_js_modules_es_object_get_own_property_descriptors__WEBPACK_IMPORTED_MODULE_9___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_object_get_own_property_descriptors__WEBPACK_IMPORTED_MODULE_9__);\n/* harmony import */ var core_js_modules_es_object_keys__WEBPACK_IMPORTED_MODULE_10__ = __webpack_require__(/*! core-js/modules/es.object.keys */ \"./node_modules/core-js/modules/es.object.keys.js\");\n/* harmony import */ var core_js_modules_es_object_keys__WEBPACK_IMPORTED_MODULE_10___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_object_keys__WEBPACK_IMPORTED_MODULE_10__);\n/* harmony import */ var core_js_modules_es_object_to_string__WEBPACK_IMPORTED_MODULE_11__ = __webpack_require__(/*! core-js/modules/es.object.to-string */ \"./node_modules/core-js/modules/es.object.to-string.js\");\n/* harmony import */ var core_js_modules_es_object_to_string__WEBPACK_IMPORTED_MODULE_11___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_object_to_string__WEBPACK_IMPORTED_MODULE_11__);\n/* harmony import */ var core_js_modules_es_regexp_to_string__WEBPACK_IMPORTED_MODULE_12__ = __webpack_require__(/*! core-js/modules/es.regexp.to-string */ \"./node_modules/core-js/modules/es.regexp.to-string.js\");\n/* harmony import */ var core_js_modules_es_regexp_to_string__WEBPACK_IMPORTED_MODULE_12___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_regexp_to_string__WEBPACK_IMPORTED_MODULE_12__);\n/* harmony import */ var core_js_modules_es_string_iterator__WEBPACK_IMPORTED_MODULE_13__ = __webpack_require__(/*! core-js/modules/es.string.iterator */ \"./node_modules/core-js/modules/es.string.iterator.js\");\n/* harmony import */ var core_js_modules_es_string_iterator__WEBPACK_IMPORTED_MODULE_13___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_string_iterator__WEBPACK_IMPORTED_MODULE_13__);\n/* harmony import */ var core_js_modules_web_dom_collections_for_each__WEBPACK_IMPORTED_MODULE_14__ = __webpack_require__(/*! core-js/modules/web.dom-collections.for-each */ \"./node_modules/core-js/modules/web.dom-collections.for-each.js\");\n/* harmony import */ var core_js_modules_web_dom_collections_for_each__WEBPACK_IMPORTED_MODULE_14___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_web_dom_collections_for_each__WEBPACK_IMPORTED_MODULE_14__);\n/* harmony import */ var core_js_modules_web_dom_collections_iterator__WEBPACK_IMPORTED_MODULE_15__ = __webpack_require__(/*! core-js/modules/web.dom-collections.iterator */ \"./node_modules/core-js/modules/web.dom-collections.iterator.js\");\n/* harmony import */ var core_js_modules_web_dom_collections_iterator__WEBPACK_IMPORTED_MODULE_15___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_web_dom_collections_iterator__WEBPACK_IMPORTED_MODULE_15__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_16__ = __webpack_require__(/*! react */ \"react\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_16___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_16__);\n/* harmony import */ var _wordpress_block_editor__WEBPACK_IMPORTED_MODULE_17__ = __webpack_require__(/*! @wordpress/block-editor */ \"@wordpress/block-editor\");\n/* harmony import */ var _wordpress_block_editor__WEBPACK_IMPORTED_MODULE_17___default = /*#__PURE__*/__webpack_require__.n(_wordpress_block_editor__WEBPACK_IMPORTED_MODULE_17__);\n/* harmony import */ var _wordpress_components__WEBPACK_IMPORTED_MODULE_18__ = __webpack_require__(/*! @wordpress/components */ \"@wordpress/components\");\n/* harmony import */ var _wordpress_components__WEBPACK_IMPORTED_MODULE_18___default = /*#__PURE__*/__webpack_require__.n(_wordpress_components__WEBPACK_IMPORTED_MODULE_18__);\n/* harmony import */ var _wordpress_data__WEBPACK_IMPORTED_MODULE_19__ = __webpack_require__(/*! @wordpress/data */ \"@wordpress/data\");\n/* harmony import */ var _wordpress_data__WEBPACK_IMPORTED_MODULE_19___default = /*#__PURE__*/__webpack_require__.n(_wordpress_data__WEBPACK_IMPORTED_MODULE_19__);\n/* harmony import */ var _wordpress_i18n__WEBPACK_IMPORTED_MODULE_20__ = __webpack_require__(/*! @wordpress/i18n */ \"@wordpress/i18n\");\n/* harmony import */ var _wordpress_i18n__WEBPACK_IMPORTED_MODULE_20___default = /*#__PURE__*/__webpack_require__.n(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_20__);\n/* harmony import */ var _hooks_use_vf_defaults__WEBPACK_IMPORTED_MODULE_21__ = __webpack_require__(/*! ../hooks/use-vf-defaults */ \"./wp-content/plugins/vf-gutenberg/blocks/hooks/use-vf-defaults.js\");\n/* harmony import */ var _vf_block_block_fields__WEBPACK_IMPORTED_MODULE_22__ = __webpack_require__(/*! ../vf-block/block-fields */ \"./wp-content/plugins/vf-gutenberg/blocks/vf-block/block-fields.jsx\");\n/* harmony import */ var _vf_grid__WEBPACK_IMPORTED_MODULE_23__ = __webpack_require__(/*! ./vf-grid */ \"./wp-content/plugins/vf-gutenberg/blocks/vf-core/vf-grid.jsx\");\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nfunction _extends() { _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; }; return _extends.apply(this, arguments); }\n\nfunction _toConsumableArray(arr) { return _arrayWithoutHoles(arr) || _iterableToArray(arr) || _nonIterableSpread(); }\n\nfunction _nonIterableSpread() { throw new TypeError(\"Invalid attempt to spread non-iterable instance\"); }\n\nfunction _iterableToArray(iter) { if (Symbol.iterator in Object(iter) || Object.prototype.toString.call(iter) === \"[object Arguments]\") return Array.from(iter); }\n\nfunction _arrayWithoutHoles(arr) { if (Array.isArray(arr)) { for (var i = 0, arr2 = new Array(arr.length); i < arr.length; i++) { arr2[i] = arr[i]; } return arr2; } }\n\nfunction ownKeys(object, enumerableOnly) { var keys = Object.keys(object); if (Object.getOwnPropertySymbols) { var symbols = Object.getOwnPropertySymbols(object); if (enumerableOnly) symbols = symbols.filter(function (sym) { return Object.getOwnPropertyDescriptor(object, sym).enumerable; }); keys.push.apply(keys, symbols); } return keys; }\n\nfunction _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i] != null ? arguments[i] : {}; if (i % 2) { ownKeys(source, true).forEach(function (key) { _defineProperty(target, key, source[key]); }); } else if (Object.getOwnPropertyDescriptors) { Object.defineProperties(target, Object.getOwnPropertyDescriptors(source)); } else { ownKeys(source).forEach(function (key) { Object.defineProperty(target, key, Object.getOwnPropertyDescriptor(source, key)); }); } } return target; }\n\nfunction _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }\n\n/**\nBlock Name: Grid\nBased on `vf-grid.jsx`\n*/\n\n\n\n\n\n\n\n\nvar defaults = Object(_hooks_use_vf_defaults__WEBPACK_IMPORTED_MODULE_21__[\"default\"])();\nvar ver = '1.0.0';\nvar MIN_COLUMNS = 2;\nvar MAX_COLUMNS = 4;\n\nvar settings = _objectSpread({}, defaults, {\n  name: 'vf/embl-grid',\n  title: Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_20__[\"__\"])('EMBL Grid'),\n  category: 'vf/core',\n  description: Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_20__[\"__\"])('Visual Framework (core)'),\n  attributes: _objectSpread({}, defaults.attributes, {\n    placeholder: {\n      type: 'integer',\n      default: 0\n    },\n    columns: {\n      type: 'integer',\n      default: 0\n    },\n    sidebar: {\n      type: 'integer',\n      default: 0\n    },\n    centered: {\n      type: 'integer',\n      default: 0\n    }\n  })\n});\n\nsettings.save = function (props) {\n  var _props$attributes = props.attributes,\n      placeholder = _props$attributes.placeholder,\n      sidebar = _props$attributes.sidebar,\n      centered = _props$attributes.centered;\n\n  if (placeholder === 1) {\n    return null;\n  }\n\n  var className = 'embl-grid';\n\n  if (!!sidebar) {\n    className += ' embl-grid--has-sidebar';\n  }\n\n  if (!!centered) {\n    className += ' embl-grid--has-centered-content';\n  }\n\n  return wp.element.createElement(\"div\", {\n    className: className\n  }, wp.element.createElement(_wordpress_block_editor__WEBPACK_IMPORTED_MODULE_17__[\"InnerBlocks\"].Content, null));\n};\n\nvar withGridDispatch = function withGridDispatch(Edit) {\n  return Object(_wordpress_data__WEBPACK_IMPORTED_MODULE_19__[\"withDispatch\"])(function (dispatch, ownProps, _ref) {\n    var select = _ref.select;\n\n    var _select = select('core/block-editor'),\n        getBlocks = _select.getBlocks;\n\n    var _dispatch = dispatch('core/block-editor'),\n        replaceInnerBlocks = _dispatch.replaceInnerBlocks; // `columns` attribute `onChange` callback\n\n\n    var setColumns = function setColumns(newColumns) {\n      var prevColumns = ownProps.attributes.columns; // Merge inner blocks when number of columns is reduced\n\n      if (newColumns < prevColumns) {\n        var columnBlocks = getBlocks(ownProps.clientId);\n        var mergeBlocks = [];\n\n        for (var i = newColumns - 1; i < prevColumns; i++) {\n          mergeBlocks.push.apply(mergeBlocks, _toConsumableArray(columnBlocks[i].innerBlocks));\n        }\n\n        replaceInnerBlocks(columnBlocks[newColumns - 1].clientId, mergeBlocks, false);\n      } // Update block attributes\n\n\n      ownProps.setAttributes({\n        placeholder: 0,\n        columns: newColumns\n      });\n\n      if (newColumns !== 3) {\n        ownProps.setAttributes({\n          sidebar: 0,\n          centered: 0\n        });\n      }\n    }; // Toggle attribute `onChange` callback\n\n\n    var setToggle = function setToggle(name, value) {\n      value = value ? 1 : 0;\n      ownProps.setAttributes(_defineProperty({\n        sidebar: 0,\n        centered: 0\n      }, name, value));\n\n      if (value) {\n        setColumns(3);\n      }\n    };\n\n    return {\n      setColumns: setColumns,\n      setToggle: setToggle\n    };\n  })(Edit);\n};\n\nsettings.edit = withGridDispatch(function (props) {\n  var _props$attributes2 = props.attributes,\n      columns = _props$attributes2.columns,\n      sidebar = _props$attributes2.sidebar,\n      centered = _props$attributes2.centered,\n      placeholder = _props$attributes2.placeholder; // ensure version is encoded in post content\n\n  props.setAttributes({\n    ver: ver\n  }); // turn on setup placeholder if no columns are defined\n\n  if (columns === 0) {\n    props.setAttributes({\n      placeholder: 1\n    });\n  } // Setup placeholder fields\n\n\n  var fields = [{\n    control: 'columns',\n    min: MIN_COLUMNS,\n    max: MAX_COLUMNS,\n    value: columns,\n    onChange: props.setColumns\n  }, {\n    label: Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_20__[\"__\"])('With Sidebar'),\n    control: 'toggle',\n    name: 'sidebar',\n    onChange: props.setToggle\n  }, {\n    label: Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_20__[\"__\"])('Centered Content'),\n    control: 'toggle',\n    name: 'centered',\n    onChange: props.setToggle\n  }]; // Return setup placeholder\n\n  if (placeholder === 1) {\n    return wp.element.createElement(\"div\", {\n      className: \"vf-block vf-block--placeholder \".concat(props.className)\n    }, wp.element.createElement(_wordpress_components__WEBPACK_IMPORTED_MODULE_18__[\"Placeholder\"], {\n      label: Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_20__[\"__\"])('EMBL Grid'),\n      icon: 'admin-generic'\n    }, wp.element.createElement(_vf_block_block_fields__WEBPACK_IMPORTED_MODULE_22__[\"default\"], _extends({}, props, {\n      fields: fields\n    }))));\n  } // Amend fields for inspector\n\n\n  fields[0].help = Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_20__[\"__\"])('Content may be reorganised when columns are reduced.');\n  fields[1].help = Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_20__[\"__\"])('3 column only.');\n  fields[2].help = fields[1].help; // Return inner blocks and inspector controls\n\n  return wp.element.createElement(react__WEBPACK_IMPORTED_MODULE_16__[\"Fragment\"], null, wp.element.createElement(_wordpress_block_editor__WEBPACK_IMPORTED_MODULE_17__[\"InspectorControls\"], null, wp.element.createElement(_wordpress_components__WEBPACK_IMPORTED_MODULE_18__[\"PanelBody\"], {\n    title: Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_20__[\"__\"])('Settings'),\n    initialOpen: true\n  }, wp.element.createElement(_vf_block_block_fields__WEBPACK_IMPORTED_MODULE_22__[\"default\"], _extends({}, props, {\n    fields: fields\n  })))), wp.element.createElement(\"div\", {\n    className: 'vf-block-grid',\n    \"data-ver\": ver,\n    \"data-embl\": true,\n    \"data-sidebar\": sidebar,\n    \"data-centered\": centered\n  }, wp.element.createElement(_wordpress_block_editor__WEBPACK_IMPORTED_MODULE_17__[\"InnerBlocks\"], {\n    allowedBlocks: ['vf/grid-column'],\n    template: Array(columns).fill(['vf/grid-column']),\n    templateLock: \"all\"\n  })));\n}); // Block transforms\n\nsettings.transforms = {\n  from: [Object(_vf_grid__WEBPACK_IMPORTED_MODULE_23__[\"fromColumns\"])('core/columns', 'vf/embl-grid', MIN_COLUMNS, MAX_COLUMNS), Object(_vf_grid__WEBPACK_IMPORTED_MODULE_23__[\"fromColumns\"])('vf/grid', 'vf/embl-grid', MIN_COLUMNS, MAX_COLUMNS)]\n};\n/* harmony default export */ __webpack_exports__[\"default\"] = (settings);\n\n//# sourceURL=webpack:///./wp-content/plugins/vf-gutenberg/blocks/vf-core/vf-embl-grid.jsx?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var core_js_modules_es_symbol__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! core-js/modules/es.symbol */ \"./node_modules/core-js/modules/es.symbol.js\");\n/* harmony import */ var core_js_modules_es_symbol__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_symbol__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var core_js_modules_es_symbol_description__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! core-js/modules/es.symbol.description */ \"./node_modules/core-js/modules/es.symbol.description.js\");\n/* harmony import */ var core_js_modules_es_symbol_description__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_symbol_description__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var core_js_modules_es_symbol_iterator__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! core-js/modules/es.symbol.iterator */ \"./node_modules/core-js/modules/es.symbol.iterator.js\");\n/* harmony import */ var core_js_modules_es_symbol_iterator__WEBPACK_IMPORTED_MODULE_2___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_symbol_iterator__WEBPACK_IMPORTED_MODULE_2__);\n/* harmony import */ var core_js_modules_es_array_fill__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! core-js/modules/es.array.fill */ \"./node_modules/core-js/modules/es.array.fill.js\");\n/* harmony import */ var core_js_modules_es_array_fill__WEBPACK_IMPORTED_MODULE_3___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_array_fill__WEBPACK_IMPORTED_MODULE_3__);\n/* harmony import */ var core_js_modules_es_array_filter__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! core-js/modules/es.array.filter */ \"./node_modules/core-js/modules/es.array.filter.js\");\n/* harmony import */ var core_js_modules_es_array_filter__WEBPACK_IMPORTED_MODULE_4___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_array_filter__WEBPACK_IMPORTED_MODULE_4__);\n/* harmony import */ var core_js_modules_es_array_from__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! core-js/modules/es.array.from */ \"./node_modules/core-js/modules/es.array.from.js\");\n/* harmony import */ var core_js_modules_es_array_from__WEBPACK_IMPORTED_MODULE_5___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_array_from__WEBPACK_IMPORTED_MODULE_5__);\n/* harmony import */ var core_js_modules_es_array_iterator__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! core-js/modules/es.array.iterator */ \"./node_modules/core-js/modules/es.array.iterator.js\");\n/* harmony import */ var core_js_modules_es_array_iterator__WEBPACK_IMPORTED_MODULE_6___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_array_iterator__WEBPACK_IMPORTED_MODULE_6__);\n/* harmony import */ var core_js_modules_es_object_assign__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! core-js/modules/es.object.assign */ \"./node_modules/core-js/modules/es.object.assign.js\");\n/* harmony import */ var core_js_modules_es_object_assign__WEBPACK_IMPORTED_MODULE_7___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_object_assign__WEBPACK_IMPORTED_MODULE_7__);\n/* harmony import */ var core_js_modules_es_object_get_own_property_descriptor__WEBPACK_IMPORTED_MODULE_8__ = __webpack_require__(/*! core-js/modules/es.object.get-own-property-descriptor */ \"./node_modules/core-js/modules/es.object.get-own-property-descriptor.js\");\n/* harmony import */ var core_js_modules_es_object_get_own_property_descriptor__WEBPACK_IMPORTED_MODULE_8___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_object_get_own_property_descriptor__WEBPACK_IMPORTED_MODULE_8__);\n/* harmony import */ var core_js_modules_es_object_get_own_property_descriptors__WEBPACK_IMPORTED_MODULE_9__ = __webpack_require__(/*! core-js/modules/es.object.get-own-property-descriptors */ \"./node_modules/core-js/modules/es.object.get-own-property-descriptors.js\");\n/* harmony import */ var core_js_modules_es_object_get_own_property_descriptors__WEBPACK_IMPORTED_MODULE_9___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_object_get_own_property_descriptors__WEBPACK_IMPORTED_MODULE_9__);\n/* harmony import */ var core_js_modules_es_object_keys__WEBPACK_IMPORTED_MODULE_10__ = __webpack_require__(/*! core-js/modules/es.object.keys */ \"./node_modules/core-js/modules/es.object.keys.js\");\n/* harmony import */ var core_js_modules_es_object_keys__WEBPACK_IMPORTED_MODULE_10___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_object_keys__WEBPACK_IMPORTED_MODULE_10__);\n/* harmony import */ var core_js_modules_es_object_to_string__WEBPACK_IMPORTED_MODULE_11__ = __webpack_require__(/*! core-js/modules/es.object.to-string */ \"./node_modules/core-js/modules/es.object.to-string.js\");\n/* harmony import */ var core_js_modules_es_object_to_string__WEBPACK_IMPORTED_MODULE_11___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_object_to_string__WEBPACK_IMPORTED_MODULE_11__);\n/* harmony import */ var core_js_modules_es_regexp_to_string__WEBPACK_IMPORTED_MODULE_12__ = __webpack_require__(/*! core-js/modules/es.regexp.to-string */ \"./node_modules/core-js/modules/es.regexp.to-string.js\");\n/* harmony import */ var core_js_modules_es_regexp_to_string__WEBPACK_IMPORTED_MODULE_12___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_regexp_to_string__WEBPACK_IMPORTED_MODULE_12__);\n/* harmony import */ var core_js_modules_es_string_iterator__WEBPACK_IMPORTED_MODULE_13__ = __webpack_require__(/*! core-js/modules/es.string.iterator */ \"./node_modules/core-js/modules/es.string.iterator.js\");\n/* harmony import */ var core_js_modules_es_string_iterator__WEBPACK_IMPORTED_MODULE_13___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_string_iterator__WEBPACK_IMPORTED_MODULE_13__);\n/* harmony import */ var core_js_modules_web_dom_collections_for_each__WEBPACK_IMPORTED_MODULE_14__ = __webpack_require__(/*! core-js/modules/web.dom-collections.for-each */ \"./node_modules/core-js/modules/web.dom-collections.for-each.js\");\n/* harmony import */ var core_js_modules_web_dom_collections_for_each__WEBPACK_IMPORTED_MODULE_14___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_web_dom_collections_for_each__WEBPACK_IMPORTED_MODULE_14__);\n/* harmony import */ var core_js_modules_web_dom_collections_iterator__WEBPACK_IMPORTED_MODULE_15__ = __webpack_require__(/*! core-js/modules/web.dom-collections.iterator */ \"./node_modules/core-js/modules/web.dom-collections.iterator.js\");\n/* harmony import */ var core_js_modules_web_dom_collections_iterator__WEBPACK_IMPORTED_MODULE_15___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_web_dom_collections_iterator__WEBPACK_IMPORTED_MODULE_15__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_16__ = __webpack_require__(/*! react */ \"react\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_16___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_16__);\n/* harmony import */ var _wordpress_block_editor__WEBPACK_IMPORTED_MODULE_17__ = __webpack_require__(/*! @wordpress/block-editor */ \"@wordpress/block-editor\");\n/* harmony import */ var _wordpress_block_editor__WEBPACK_IMPORTED_MODULE_17___default = /*#__PURE__*/__webpack_require__.n(_wordpress_block_editor__WEBPACK_IMPORTED_MODULE_17__);\n/* harmony import */ var _wordpress_components__WEBPACK_IMPORTED_MODULE_18__ = __webpack_require__(/*! @wordpress/components */ \"@wordpress/components\");\n/* harmony import */ var _wordpress_components__WEBPACK_IMPORTED_MODULE_18___default = /*#__PURE__*/__webpack_require__.n(_wordpress_components__WEBPACK_IMPORTED_MODULE_18__);\n/* harmony import */ var _wordpress_data__WEBPACK_IMPORTED_MODULE_19__ = __webpack_require__(/*! @wordpress/data */ \"@wordpress/data\");\n/* harmony import */ var _wordpress_data__WEBPACK_IMPORTED_MODULE_19___default = /*#__PURE__*/__webpack_require__.n(_wordpress_data__WEBPACK_IMPORTED_MODULE_19__);\n/* harmony import */ var _wordpress_i18n__WEBPACK_IMPORTED_MODULE_20__ = __webpack_require__(/*! @wordpress/i18n */ \"@wordpress/i18n\");\n/* harmony import */ var _wordpress_i18n__WEBPACK_IMPORTED_MODULE_20___default = /*#__PURE__*/__webpack_require__.n(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_20__);\n/* harmony import */ var _hooks_use_vf_defaults__WEBPACK_IMPORTED_MODULE_21__ = __webpack_require__(/*! ../hooks/use-vf-defaults */ \"./wp-content/plugins/vf-gutenberg/blocks/hooks/use-vf-defaults.js\");\n/* harmony import */ var _vf_block_block_fields__WEBPACK_IMPORTED_MODULE_22__ = __webpack_require__(/*! ../vf-block/block-fields */ \"./wp-content/plugins/vf-gutenberg/blocks/vf-block/block-fields.jsx\");\n/* harmony import */ var _transforms_grid__WEBPACK_IMPORTED_MODULE_23__ = __webpack_require__(/*! ./transforms/grid */ \"./wp-content/plugins/vf-gutenberg/blocks/vf-core/transforms/grid.js\");\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nfunction _extends() { _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; }; return _extends.apply(this, arguments); }\n\nfunction _toConsumableArray(arr) { return _arrayWithoutHoles(arr) || _iterableToArray(arr) || _nonIterableSpread(); }\n\nfunction _nonIterableSpread() { throw new TypeError(\"Invalid attempt to spread non-iterable instance\"); }\n\nfunction _iterableToArray(iter) { if (Symbol.iterator in Object(iter) || Object.prototype.toString.call(iter) === \"[object Arguments]\") return Array.from(iter); }\n\nfunction _arrayWithoutHoles(arr) { if (Array.isArray(arr)) { for (var i = 0, arr2 = new Array(arr.length); i < arr.length; i++) { arr2[i] = arr[i]; } return arr2; } }\n\nfunction ownKeys(object, enumerableOnly) { var keys = Object.keys(object); if (Object.getOwnPropertySymbols) { var symbols = Object.getOwnPropertySymbols(object); if (enumerableOnly) symbols = symbols.filter(function (sym) { return Object.getOwnPropertyDescriptor(object, sym).enumerable; }); keys.push.apply(keys, symbols); } return keys; }\n\nfunction _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i] != null ? arguments[i] : {}; if (i % 2) { ownKeys(source, true).forEach(function (key) { _defineProperty(target, key, source[key]); }); } else if (Object.getOwnPropertyDescriptors) { Object.defineProperties(target, Object.getOwnPropertyDescriptors(source)); } else { ownKeys(source).forEach(function (key) { Object.defineProperty(target, key, Object.getOwnPropertyDescriptor(source, key)); }); } } return target; }\n\nfunction _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }\n\n/**\nBlock Name: Grid\nBased on `vf-grid.jsx`\n*/\n\n\n\n\n\n\n\n\nvar defaults = Object(_hooks_use_vf_defaults__WEBPACK_IMPORTED_MODULE_21__[\"default\"])();\nvar ver = '1.0.0';\nvar MIN_COLUMNS = 2;\nvar MAX_COLUMNS = 4;\n\nvar settings = _objectSpread({}, defaults, {\n  name: 'vf/embl-grid',\n  title: Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_20__[\"__\"])('EMBL Grid'),\n  category: 'vf/core',\n  description: Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_20__[\"__\"])('Visual Framework (core)'),\n  attributes: _objectSpread({}, defaults.attributes, {\n    placeholder: {\n      type: 'integer',\n      default: 0\n    },\n    columns: {\n      type: 'integer',\n      default: 0\n    },\n    sidebar: {\n      type: 'integer',\n      default: 0\n    },\n    centered: {\n      type: 'integer',\n      default: 0\n    }\n  })\n});\n\nsettings.save = function (props) {\n  var _props$attributes = props.attributes,\n      placeholder = _props$attributes.placeholder,\n      sidebar = _props$attributes.sidebar,\n      centered = _props$attributes.centered;\n\n  if (placeholder === 1) {\n    return null;\n  }\n\n  var className = 'embl-grid';\n\n  if (!!sidebar) {\n    className += ' embl-grid--has-sidebar';\n  }\n\n  if (!!centered) {\n    className += ' embl-grid--has-centered-content';\n  }\n\n  return wp.element.createElement(\"div\", {\n    className: className\n  }, wp.element.createElement(_wordpress_block_editor__WEBPACK_IMPORTED_MODULE_17__[\"InnerBlocks\"].Content, null));\n};\n\nvar withGridDispatch = function withGridDispatch(Edit) {\n  return Object(_wordpress_data__WEBPACK_IMPORTED_MODULE_19__[\"withDispatch\"])(function (dispatch, ownProps, _ref) {\n    var select = _ref.select;\n\n    var _select = select('core/block-editor'),\n        getBlocks = _select.getBlocks;\n\n    var _dispatch = dispatch('core/block-editor'),\n        replaceInnerBlocks = _dispatch.replaceInnerBlocks; // `columns` attribute `onChange` callback\n\n\n    var setColumns = function setColumns(newColumns) {\n      var prevColumns = ownProps.attributes.columns; // Merge inner blocks when number of columns is reduced\n\n      if (newColumns < prevColumns) {\n        var columnBlocks = getBlocks(ownProps.clientId);\n        var mergeBlocks = [];\n\n        for (var i = newColumns - 1; i < prevColumns; i++) {\n          mergeBlocks.push.apply(mergeBlocks, _toConsumableArray(columnBlocks[i].innerBlocks));\n        }\n\n        replaceInnerBlocks(columnBlocks[newColumns - 1].clientId, mergeBlocks, false);\n      } // Update block attributes\n\n\n      ownProps.setAttributes({\n        placeholder: 0,\n        columns: newColumns\n      });\n\n      if (newColumns !== 3) {\n        ownProps.setAttributes({\n          sidebar: 0,\n          centered: 0\n        });\n      }\n    }; // Toggle attribute `onChange` callback\n\n\n    var setToggle = function setToggle(name, value) {\n      value = value ? 1 : 0;\n      ownProps.setAttributes(_defineProperty({\n        sidebar: 0,\n        centered: 0\n      }, name, value));\n\n      if (value) {\n        setColumns(3);\n      }\n    };\n\n    return {\n      setColumns: setColumns,\n      setToggle: setToggle\n    };\n  })(Edit);\n};\n\nsettings.edit = withGridDispatch(function (props) {\n  var _props$attributes2 = props.attributes,\n      columns = _props$attributes2.columns,\n      sidebar = _props$attributes2.sidebar,\n      centered = _props$attributes2.centered,\n      placeholder = _props$attributes2.placeholder; // ensure version is encoded in post content\n\n  props.setAttributes({\n    ver: ver\n  }); // turn on setup placeholder if no columns are defined\n\n  if (columns === 0) {\n    props.setAttributes({\n      placeholder: 1\n    });\n  } // Setup placeholder fields\n\n\n  var fields = [{\n    control: 'columns',\n    min: MIN_COLUMNS,\n    max: MAX_COLUMNS,\n    value: columns,\n    onChange: props.setColumns\n  }, {\n    label: Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_20__[\"__\"])('With Sidebar'),\n    control: 'toggle',\n    name: 'sidebar',\n    onChange: props.setToggle\n  }, {\n    label: Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_20__[\"__\"])('Centered Content'),\n    control: 'toggle',\n    name: 'centered',\n    onChange: props.setToggle\n  }]; // Return setup placeholder\n\n  if (placeholder === 1) {\n    return wp.element.createElement(\"div\", {\n      className: \"vf-block vf-block--placeholder \".concat(props.className)\n    }, wp.element.createElement(_wordpress_components__WEBPACK_IMPORTED_MODULE_18__[\"Placeholder\"], {\n      label: Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_20__[\"__\"])('EMBL Grid'),\n      icon: 'admin-generic'\n    }, wp.element.createElement(_vf_block_block_fields__WEBPACK_IMPORTED_MODULE_22__[\"default\"], _extends({}, props, {\n      fields: fields\n    }))));\n  } // Amend fields for inspector\n\n\n  fields[0].help = Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_20__[\"__\"])('Content may be reorganised when columns are reduced.');\n  fields[1].help = Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_20__[\"__\"])('3 column only.');\n  fields[2].help = fields[1].help; // Return inner blocks and inspector controls\n\n  return wp.element.createElement(react__WEBPACK_IMPORTED_MODULE_16__[\"Fragment\"], null, wp.element.createElement(_wordpress_block_editor__WEBPACK_IMPORTED_MODULE_17__[\"InspectorControls\"], null, wp.element.createElement(_wordpress_components__WEBPACK_IMPORTED_MODULE_18__[\"PanelBody\"], {\n    title: Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_20__[\"__\"])('Settings'),\n    initialOpen: true\n  }, wp.element.createElement(_vf_block_block_fields__WEBPACK_IMPORTED_MODULE_22__[\"default\"], _extends({}, props, {\n    fields: fields\n  })))), wp.element.createElement(\"div\", {\n    className: 'vf-block-grid',\n    \"data-ver\": ver,\n    \"data-embl\": true,\n    \"data-sidebar\": sidebar,\n    \"data-centered\": centered\n  }, wp.element.createElement(_wordpress_block_editor__WEBPACK_IMPORTED_MODULE_17__[\"InnerBlocks\"], {\n    allowedBlocks: ['vf/grid-column'],\n    template: Array(columns).fill(['vf/grid-column']),\n    templateLock: \"all\"\n  })));\n}); // Block transforms\n\nsettings.transforms = {\n  from: [Object(_transforms_grid__WEBPACK_IMPORTED_MODULE_23__[\"fromColumns\"])('core/columns', 'vf/embl-grid', MIN_COLUMNS, MAX_COLUMNS), Object(_transforms_grid__WEBPACK_IMPORTED_MODULE_23__[\"fromColumns\"])('vf/grid', 'vf/embl-grid', MIN_COLUMNS, MAX_COLUMNS)]\n};\n/* harmony default export */ __webpack_exports__[\"default\"] = (settings);\n\n//# sourceURL=webpack:///./wp-content/plugins/vf-gutenberg/blocks/vf-core/vf-embl-grid.jsx?");
 
 /***/ }),
 
@@ -1932,11 +1944,11 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var core
 /*!********************************************************************!*\
   !*** ./wp-content/plugins/vf-gutenberg/blocks/vf-core/vf-grid.jsx ***!
   \********************************************************************/
-/*! exports provided: fromColumns, default */
+/*! exports provided: default */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"fromColumns\", function() { return fromColumns; });\n/* harmony import */ var core_js_modules_es_symbol__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! core-js/modules/es.symbol */ \"./node_modules/core-js/modules/es.symbol.js\");\n/* harmony import */ var core_js_modules_es_symbol__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_symbol__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var core_js_modules_es_symbol_description__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! core-js/modules/es.symbol.description */ \"./node_modules/core-js/modules/es.symbol.description.js\");\n/* harmony import */ var core_js_modules_es_symbol_description__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_symbol_description__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var core_js_modules_es_symbol_iterator__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! core-js/modules/es.symbol.iterator */ \"./node_modules/core-js/modules/es.symbol.iterator.js\");\n/* harmony import */ var core_js_modules_es_symbol_iterator__WEBPACK_IMPORTED_MODULE_2___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_symbol_iterator__WEBPACK_IMPORTED_MODULE_2__);\n/* harmony import */ var core_js_modules_es_array_fill__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! core-js/modules/es.array.fill */ \"./node_modules/core-js/modules/es.array.fill.js\");\n/* harmony import */ var core_js_modules_es_array_fill__WEBPACK_IMPORTED_MODULE_3___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_array_fill__WEBPACK_IMPORTED_MODULE_3__);\n/* harmony import */ var core_js_modules_es_array_filter__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! core-js/modules/es.array.filter */ \"./node_modules/core-js/modules/es.array.filter.js\");\n/* harmony import */ var core_js_modules_es_array_filter__WEBPACK_IMPORTED_MODULE_4___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_array_filter__WEBPACK_IMPORTED_MODULE_4__);\n/* harmony import */ var core_js_modules_es_array_from__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! core-js/modules/es.array.from */ \"./node_modules/core-js/modules/es.array.from.js\");\n/* harmony import */ var core_js_modules_es_array_from__WEBPACK_IMPORTED_MODULE_5___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_array_from__WEBPACK_IMPORTED_MODULE_5__);\n/* harmony import */ var core_js_modules_es_array_iterator__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! core-js/modules/es.array.iterator */ \"./node_modules/core-js/modules/es.array.iterator.js\");\n/* harmony import */ var core_js_modules_es_array_iterator__WEBPACK_IMPORTED_MODULE_6___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_array_iterator__WEBPACK_IMPORTED_MODULE_6__);\n/* harmony import */ var core_js_modules_es_array_map__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! core-js/modules/es.array.map */ \"./node_modules/core-js/modules/es.array.map.js\");\n/* harmony import */ var core_js_modules_es_array_map__WEBPACK_IMPORTED_MODULE_7___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_array_map__WEBPACK_IMPORTED_MODULE_7__);\n/* harmony import */ var core_js_modules_es_object_get_own_property_descriptor__WEBPACK_IMPORTED_MODULE_8__ = __webpack_require__(/*! core-js/modules/es.object.get-own-property-descriptor */ \"./node_modules/core-js/modules/es.object.get-own-property-descriptor.js\");\n/* harmony import */ var core_js_modules_es_object_get_own_property_descriptor__WEBPACK_IMPORTED_MODULE_8___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_object_get_own_property_descriptor__WEBPACK_IMPORTED_MODULE_8__);\n/* harmony import */ var core_js_modules_es_object_get_own_property_descriptors__WEBPACK_IMPORTED_MODULE_9__ = __webpack_require__(/*! core-js/modules/es.object.get-own-property-descriptors */ \"./node_modules/core-js/modules/es.object.get-own-property-descriptors.js\");\n/* harmony import */ var core_js_modules_es_object_get_own_property_descriptors__WEBPACK_IMPORTED_MODULE_9___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_object_get_own_property_descriptors__WEBPACK_IMPORTED_MODULE_9__);\n/* harmony import */ var core_js_modules_es_object_keys__WEBPACK_IMPORTED_MODULE_10__ = __webpack_require__(/*! core-js/modules/es.object.keys */ \"./node_modules/core-js/modules/es.object.keys.js\");\n/* harmony import */ var core_js_modules_es_object_keys__WEBPACK_IMPORTED_MODULE_10___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_object_keys__WEBPACK_IMPORTED_MODULE_10__);\n/* harmony import */ var core_js_modules_es_object_to_string__WEBPACK_IMPORTED_MODULE_11__ = __webpack_require__(/*! core-js/modules/es.object.to-string */ \"./node_modules/core-js/modules/es.object.to-string.js\");\n/* harmony import */ var core_js_modules_es_object_to_string__WEBPACK_IMPORTED_MODULE_11___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_object_to_string__WEBPACK_IMPORTED_MODULE_11__);\n/* harmony import */ var core_js_modules_es_regexp_to_string__WEBPACK_IMPORTED_MODULE_12__ = __webpack_require__(/*! core-js/modules/es.regexp.to-string */ \"./node_modules/core-js/modules/es.regexp.to-string.js\");\n/* harmony import */ var core_js_modules_es_regexp_to_string__WEBPACK_IMPORTED_MODULE_12___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_regexp_to_string__WEBPACK_IMPORTED_MODULE_12__);\n/* harmony import */ var core_js_modules_es_string_iterator__WEBPACK_IMPORTED_MODULE_13__ = __webpack_require__(/*! core-js/modules/es.string.iterator */ \"./node_modules/core-js/modules/es.string.iterator.js\");\n/* harmony import */ var core_js_modules_es_string_iterator__WEBPACK_IMPORTED_MODULE_13___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_string_iterator__WEBPACK_IMPORTED_MODULE_13__);\n/* harmony import */ var core_js_modules_web_dom_collections_for_each__WEBPACK_IMPORTED_MODULE_14__ = __webpack_require__(/*! core-js/modules/web.dom-collections.for-each */ \"./node_modules/core-js/modules/web.dom-collections.for-each.js\");\n/* harmony import */ var core_js_modules_web_dom_collections_for_each__WEBPACK_IMPORTED_MODULE_14___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_web_dom_collections_for_each__WEBPACK_IMPORTED_MODULE_14__);\n/* harmony import */ var core_js_modules_web_dom_collections_iterator__WEBPACK_IMPORTED_MODULE_15__ = __webpack_require__(/*! core-js/modules/web.dom-collections.iterator */ \"./node_modules/core-js/modules/web.dom-collections.iterator.js\");\n/* harmony import */ var core_js_modules_web_dom_collections_iterator__WEBPACK_IMPORTED_MODULE_15___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_web_dom_collections_iterator__WEBPACK_IMPORTED_MODULE_15__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_16__ = __webpack_require__(/*! react */ \"react\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_16___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_16__);\n/* harmony import */ var _wordpress_blocks__WEBPACK_IMPORTED_MODULE_17__ = __webpack_require__(/*! @wordpress/blocks */ \"@wordpress/blocks\");\n/* harmony import */ var _wordpress_blocks__WEBPACK_IMPORTED_MODULE_17___default = /*#__PURE__*/__webpack_require__.n(_wordpress_blocks__WEBPACK_IMPORTED_MODULE_17__);\n/* harmony import */ var _wordpress_block_editor__WEBPACK_IMPORTED_MODULE_18__ = __webpack_require__(/*! @wordpress/block-editor */ \"@wordpress/block-editor\");\n/* harmony import */ var _wordpress_block_editor__WEBPACK_IMPORTED_MODULE_18___default = /*#__PURE__*/__webpack_require__.n(_wordpress_block_editor__WEBPACK_IMPORTED_MODULE_18__);\n/* harmony import */ var _wordpress_components__WEBPACK_IMPORTED_MODULE_19__ = __webpack_require__(/*! @wordpress/components */ \"@wordpress/components\");\n/* harmony import */ var _wordpress_components__WEBPACK_IMPORTED_MODULE_19___default = /*#__PURE__*/__webpack_require__.n(_wordpress_components__WEBPACK_IMPORTED_MODULE_19__);\n/* harmony import */ var _wordpress_data__WEBPACK_IMPORTED_MODULE_20__ = __webpack_require__(/*! @wordpress/data */ \"@wordpress/data\");\n/* harmony import */ var _wordpress_data__WEBPACK_IMPORTED_MODULE_20___default = /*#__PURE__*/__webpack_require__.n(_wordpress_data__WEBPACK_IMPORTED_MODULE_20__);\n/* harmony import */ var _wordpress_i18n__WEBPACK_IMPORTED_MODULE_21__ = __webpack_require__(/*! @wordpress/i18n */ \"@wordpress/i18n\");\n/* harmony import */ var _wordpress_i18n__WEBPACK_IMPORTED_MODULE_21___default = /*#__PURE__*/__webpack_require__.n(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_21__);\n/* harmony import */ var _hooks_use_vf_defaults__WEBPACK_IMPORTED_MODULE_22__ = __webpack_require__(/*! ../hooks/use-vf-defaults */ \"./wp-content/plugins/vf-gutenberg/blocks/hooks/use-vf-defaults.js\");\n/* harmony import */ var _vf_block_block_fields__WEBPACK_IMPORTED_MODULE_23__ = __webpack_require__(/*! ../vf-block/block-fields */ \"./wp-content/plugins/vf-gutenberg/blocks/vf-block/block-fields.jsx\");\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nfunction _toConsumableArray(arr) { return _arrayWithoutHoles(arr) || _iterableToArray(arr) || _nonIterableSpread(); }\n\nfunction _nonIterableSpread() { throw new TypeError(\"Invalid attempt to spread non-iterable instance\"); }\n\nfunction _iterableToArray(iter) { if (Symbol.iterator in Object(iter) || Object.prototype.toString.call(iter) === \"[object Arguments]\") return Array.from(iter); }\n\nfunction _arrayWithoutHoles(arr) { if (Array.isArray(arr)) { for (var i = 0, arr2 = new Array(arr.length); i < arr.length; i++) { arr2[i] = arr[i]; } return arr2; } }\n\nfunction ownKeys(object, enumerableOnly) { var keys = Object.keys(object); if (Object.getOwnPropertySymbols) { var symbols = Object.getOwnPropertySymbols(object); if (enumerableOnly) symbols = symbols.filter(function (sym) { return Object.getOwnPropertyDescriptor(object, sym).enumerable; }); keys.push.apply(keys, symbols); } return keys; }\n\nfunction _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i] != null ? arguments[i] : {}; if (i % 2) { ownKeys(source, true).forEach(function (key) { _defineProperty(target, key, source[key]); }); } else if (Object.getOwnPropertyDescriptors) { Object.defineProperties(target, Object.getOwnPropertyDescriptors(source)); } else { ownKeys(source).forEach(function (key) { Object.defineProperty(target, key, Object.getOwnPropertyDescriptor(source, key)); }); } } return target; }\n\nfunction _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }\n\n/**\nBlock Name: Grid\n*/\n\n\n\n\n\n\n\n // Return a transform object for grids using `vf/grid-column`\n\nvar fromColumns = function fromColumns(fromBlock, toBlock, min, max) {\n  return {\n    type: 'block',\n    blocks: [fromBlock],\n    transform: function transform(attributes, innerBlocks) {\n      // Map column props\n      var innerProps = innerBlocks.map(function (block) {\n        return {\n          attributes: _objectSpread({}, block.attributes),\n          innerBlocks: _toConsumableArray(block.innerBlocks)\n        };\n      }); // Fill empty props to match min number of columns\n\n      while (innerProps.length < min) {\n        innerProps.push({});\n      } // Merge end props to match max number of columns\n\n\n      while (innerProps.length > max) {\n        var _innerProps$innerBloc;\n\n        var mergeProps = innerProps.pop();\n\n        (_innerProps$innerBloc = innerProps[innerProps.length - 1].innerBlocks).push.apply(_innerProps$innerBloc, _toConsumableArray(mergeProps.innerBlocks));\n      } // Return new grid block with inner columns\n\n\n      return Object(_wordpress_blocks__WEBPACK_IMPORTED_MODULE_17__[\"createBlock\"])(toBlock, {\n        columns: innerProps.length\n      }, innerProps.map(function (props) {\n        return Object(_wordpress_blocks__WEBPACK_IMPORTED_MODULE_17__[\"createBlock\"])('vf/grid-column', props.attributes || {}, props.innerBlocks || []);\n      }));\n    }\n  };\n};\nvar defaults = Object(_hooks_use_vf_defaults__WEBPACK_IMPORTED_MODULE_22__[\"default\"])();\nvar ver = '1.0.0';\nvar MIN_COLUMNS = 1;\nvar MAX_COLUMNS = 6;\n\nvar settings = _objectSpread({}, defaults, {\n  name: 'vf/grid',\n  title: Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_21__[\"__\"])('VF Grid'),\n  category: 'vf/core',\n  description: Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_21__[\"__\"])('Visual Framework (core)'),\n  attributes: _objectSpread({}, defaults.attributes, {\n    placeholder: {\n      type: 'integer',\n      default: 0\n    },\n    columns: {\n      type: 'integer',\n      default: 0\n    }\n  })\n});\n\nsettings.save = function (props) {\n  if (props.attributes.placeholder === 1) {\n    return null;\n  }\n\n  var columns = props.attributes.columns;\n  var className = \"vf-grid | vf-grid__col-\".concat(columns);\n  return wp.element.createElement(\"div\", {\n    className: className\n  }, wp.element.createElement(_wordpress_block_editor__WEBPACK_IMPORTED_MODULE_18__[\"InnerBlocks\"].Content, null));\n};\n\nvar withGridDispatch = function withGridDispatch(Edit) {\n  return Object(_wordpress_data__WEBPACK_IMPORTED_MODULE_20__[\"withDispatch\"])(function (dispatch, ownProps, _ref) {\n    var select = _ref.select;\n\n    var _select = select('core/block-editor'),\n        getBlocks = _select.getBlocks;\n\n    var _dispatch = dispatch('core/block-editor'),\n        replaceInnerBlocks = _dispatch.replaceInnerBlocks; // `columns` attribute `onChange` callback\n\n\n    var setColumns = function setColumns(newColumns) {\n      var prevColumns = ownProps.attributes.columns; // Merge inner blocks when number of columns is reduced\n\n      if (newColumns < prevColumns) {\n        var columnBlocks = getBlocks(ownProps.clientId);\n        var mergeBlocks = [];\n\n        for (var i = newColumns - 1; i < prevColumns; i++) {\n          mergeBlocks.push.apply(mergeBlocks, _toConsumableArray(columnBlocks[i].innerBlocks));\n        }\n\n        replaceInnerBlocks(columnBlocks[newColumns - 1].clientId, mergeBlocks, false);\n      } // Update block attributes\n\n\n      ownProps.setAttributes({\n        columns: newColumns,\n        placeholder: 0\n      });\n    };\n\n    return {\n      setColumns: setColumns\n    };\n  })(Edit);\n};\n\nsettings.edit = withGridDispatch(function (props) {\n  var _props$attributes = props.attributes,\n      columns = _props$attributes.columns,\n      placeholder = _props$attributes.placeholder; // Ensure version is encoded in post content\n\n  props.setAttributes({\n    ver: ver\n  }); // Turn on setup placeholder if no columns are defined\n\n  if (columns === 0) {\n    props.setAttributes({\n      placeholder: 1\n    });\n  } // Setup placeholder fields\n\n\n  var fields = [{\n    control: 'columns',\n    min: MIN_COLUMNS,\n    max: MAX_COLUMNS,\n    value: columns,\n    onChange: props.setColumns\n  }]; // Return setup placeholder\n\n  if (placeholder === 1) {\n    return wp.element.createElement(\"div\", {\n      className: \"vf-block vf-block--placeholder \".concat(props.className)\n    }, wp.element.createElement(_wordpress_components__WEBPACK_IMPORTED_MODULE_19__[\"Placeholder\"], {\n      label: Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_21__[\"__\"])('VF Grid'),\n      icon: 'admin-generic'\n    }, wp.element.createElement(_vf_block_block_fields__WEBPACK_IMPORTED_MODULE_23__[\"default\"], {\n      fields: fields\n    })));\n  } // Amend fields for inspector\n\n\n  fields[0].help = Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_21__[\"__\"])('Content may be reorganised when columns are reduced.'); // Return inner blocks and inspector controls\n\n  return wp.element.createElement(react__WEBPACK_IMPORTED_MODULE_16__[\"Fragment\"], null, wp.element.createElement(_wordpress_block_editor__WEBPACK_IMPORTED_MODULE_18__[\"InspectorControls\"], null, wp.element.createElement(_wordpress_components__WEBPACK_IMPORTED_MODULE_19__[\"PanelBody\"], {\n    title: Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_21__[\"__\"])('Settings'),\n    initialOpen: true\n  }, wp.element.createElement(_vf_block_block_fields__WEBPACK_IMPORTED_MODULE_23__[\"default\"], {\n    fields: fields\n  }))), wp.element.createElement(\"div\", {\n    className: 'vf-block-grid',\n    \"data-ver\": ver,\n    \"data-columns\": columns\n  }, wp.element.createElement(_wordpress_block_editor__WEBPACK_IMPORTED_MODULE_18__[\"InnerBlocks\"], {\n    allowedBlocks: ['vf/grid-column'],\n    template: Array(columns).fill(['vf/grid-column']),\n    templateLock: \"all\"\n  })));\n}); // Block transforms\n\nsettings.transforms = {\n  from: [fromColumns('core/columns', 'vf/grid', MIN_COLUMNS, MAX_COLUMNS), fromColumns('vf/embl-grid', 'vf/grid', MIN_COLUMNS, MAX_COLUMNS)]\n};\n/* harmony default export */ __webpack_exports__[\"default\"] = (settings);\n\n//# sourceURL=webpack:///./wp-content/plugins/vf-gutenberg/blocks/vf-core/vf-grid.jsx?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var core_js_modules_es_symbol__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! core-js/modules/es.symbol */ \"./node_modules/core-js/modules/es.symbol.js\");\n/* harmony import */ var core_js_modules_es_symbol__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_symbol__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var core_js_modules_es_symbol_description__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! core-js/modules/es.symbol.description */ \"./node_modules/core-js/modules/es.symbol.description.js\");\n/* harmony import */ var core_js_modules_es_symbol_description__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_symbol_description__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var core_js_modules_es_symbol_iterator__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! core-js/modules/es.symbol.iterator */ \"./node_modules/core-js/modules/es.symbol.iterator.js\");\n/* harmony import */ var core_js_modules_es_symbol_iterator__WEBPACK_IMPORTED_MODULE_2___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_symbol_iterator__WEBPACK_IMPORTED_MODULE_2__);\n/* harmony import */ var core_js_modules_es_array_fill__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! core-js/modules/es.array.fill */ \"./node_modules/core-js/modules/es.array.fill.js\");\n/* harmony import */ var core_js_modules_es_array_fill__WEBPACK_IMPORTED_MODULE_3___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_array_fill__WEBPACK_IMPORTED_MODULE_3__);\n/* harmony import */ var core_js_modules_es_array_filter__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! core-js/modules/es.array.filter */ \"./node_modules/core-js/modules/es.array.filter.js\");\n/* harmony import */ var core_js_modules_es_array_filter__WEBPACK_IMPORTED_MODULE_4___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_array_filter__WEBPACK_IMPORTED_MODULE_4__);\n/* harmony import */ var core_js_modules_es_array_from__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! core-js/modules/es.array.from */ \"./node_modules/core-js/modules/es.array.from.js\");\n/* harmony import */ var core_js_modules_es_array_from__WEBPACK_IMPORTED_MODULE_5___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_array_from__WEBPACK_IMPORTED_MODULE_5__);\n/* harmony import */ var core_js_modules_es_array_iterator__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! core-js/modules/es.array.iterator */ \"./node_modules/core-js/modules/es.array.iterator.js\");\n/* harmony import */ var core_js_modules_es_array_iterator__WEBPACK_IMPORTED_MODULE_6___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_array_iterator__WEBPACK_IMPORTED_MODULE_6__);\n/* harmony import */ var core_js_modules_es_object_get_own_property_descriptor__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! core-js/modules/es.object.get-own-property-descriptor */ \"./node_modules/core-js/modules/es.object.get-own-property-descriptor.js\");\n/* harmony import */ var core_js_modules_es_object_get_own_property_descriptor__WEBPACK_IMPORTED_MODULE_7___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_object_get_own_property_descriptor__WEBPACK_IMPORTED_MODULE_7__);\n/* harmony import */ var core_js_modules_es_object_get_own_property_descriptors__WEBPACK_IMPORTED_MODULE_8__ = __webpack_require__(/*! core-js/modules/es.object.get-own-property-descriptors */ \"./node_modules/core-js/modules/es.object.get-own-property-descriptors.js\");\n/* harmony import */ var core_js_modules_es_object_get_own_property_descriptors__WEBPACK_IMPORTED_MODULE_8___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_object_get_own_property_descriptors__WEBPACK_IMPORTED_MODULE_8__);\n/* harmony import */ var core_js_modules_es_object_keys__WEBPACK_IMPORTED_MODULE_9__ = __webpack_require__(/*! core-js/modules/es.object.keys */ \"./node_modules/core-js/modules/es.object.keys.js\");\n/* harmony import */ var core_js_modules_es_object_keys__WEBPACK_IMPORTED_MODULE_9___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_object_keys__WEBPACK_IMPORTED_MODULE_9__);\n/* harmony import */ var core_js_modules_es_object_to_string__WEBPACK_IMPORTED_MODULE_10__ = __webpack_require__(/*! core-js/modules/es.object.to-string */ \"./node_modules/core-js/modules/es.object.to-string.js\");\n/* harmony import */ var core_js_modules_es_object_to_string__WEBPACK_IMPORTED_MODULE_10___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_object_to_string__WEBPACK_IMPORTED_MODULE_10__);\n/* harmony import */ var core_js_modules_es_regexp_to_string__WEBPACK_IMPORTED_MODULE_11__ = __webpack_require__(/*! core-js/modules/es.regexp.to-string */ \"./node_modules/core-js/modules/es.regexp.to-string.js\");\n/* harmony import */ var core_js_modules_es_regexp_to_string__WEBPACK_IMPORTED_MODULE_11___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_regexp_to_string__WEBPACK_IMPORTED_MODULE_11__);\n/* harmony import */ var core_js_modules_es_string_iterator__WEBPACK_IMPORTED_MODULE_12__ = __webpack_require__(/*! core-js/modules/es.string.iterator */ \"./node_modules/core-js/modules/es.string.iterator.js\");\n/* harmony import */ var core_js_modules_es_string_iterator__WEBPACK_IMPORTED_MODULE_12___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_es_string_iterator__WEBPACK_IMPORTED_MODULE_12__);\n/* harmony import */ var core_js_modules_web_dom_collections_for_each__WEBPACK_IMPORTED_MODULE_13__ = __webpack_require__(/*! core-js/modules/web.dom-collections.for-each */ \"./node_modules/core-js/modules/web.dom-collections.for-each.js\");\n/* harmony import */ var core_js_modules_web_dom_collections_for_each__WEBPACK_IMPORTED_MODULE_13___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_web_dom_collections_for_each__WEBPACK_IMPORTED_MODULE_13__);\n/* harmony import */ var core_js_modules_web_dom_collections_iterator__WEBPACK_IMPORTED_MODULE_14__ = __webpack_require__(/*! core-js/modules/web.dom-collections.iterator */ \"./node_modules/core-js/modules/web.dom-collections.iterator.js\");\n/* harmony import */ var core_js_modules_web_dom_collections_iterator__WEBPACK_IMPORTED_MODULE_14___default = /*#__PURE__*/__webpack_require__.n(core_js_modules_web_dom_collections_iterator__WEBPACK_IMPORTED_MODULE_14__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_15__ = __webpack_require__(/*! react */ \"react\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_15___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_15__);\n/* harmony import */ var _wordpress_block_editor__WEBPACK_IMPORTED_MODULE_16__ = __webpack_require__(/*! @wordpress/block-editor */ \"@wordpress/block-editor\");\n/* harmony import */ var _wordpress_block_editor__WEBPACK_IMPORTED_MODULE_16___default = /*#__PURE__*/__webpack_require__.n(_wordpress_block_editor__WEBPACK_IMPORTED_MODULE_16__);\n/* harmony import */ var _wordpress_components__WEBPACK_IMPORTED_MODULE_17__ = __webpack_require__(/*! @wordpress/components */ \"@wordpress/components\");\n/* harmony import */ var _wordpress_components__WEBPACK_IMPORTED_MODULE_17___default = /*#__PURE__*/__webpack_require__.n(_wordpress_components__WEBPACK_IMPORTED_MODULE_17__);\n/* harmony import */ var _wordpress_data__WEBPACK_IMPORTED_MODULE_18__ = __webpack_require__(/*! @wordpress/data */ \"@wordpress/data\");\n/* harmony import */ var _wordpress_data__WEBPACK_IMPORTED_MODULE_18___default = /*#__PURE__*/__webpack_require__.n(_wordpress_data__WEBPACK_IMPORTED_MODULE_18__);\n/* harmony import */ var _wordpress_i18n__WEBPACK_IMPORTED_MODULE_19__ = __webpack_require__(/*! @wordpress/i18n */ \"@wordpress/i18n\");\n/* harmony import */ var _wordpress_i18n__WEBPACK_IMPORTED_MODULE_19___default = /*#__PURE__*/__webpack_require__.n(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_19__);\n/* harmony import */ var _hooks_use_vf_defaults__WEBPACK_IMPORTED_MODULE_20__ = __webpack_require__(/*! ../hooks/use-vf-defaults */ \"./wp-content/plugins/vf-gutenberg/blocks/hooks/use-vf-defaults.js\");\n/* harmony import */ var _vf_block_block_fields__WEBPACK_IMPORTED_MODULE_21__ = __webpack_require__(/*! ../vf-block/block-fields */ \"./wp-content/plugins/vf-gutenberg/blocks/vf-block/block-fields.jsx\");\n/* harmony import */ var _transforms_grid__WEBPACK_IMPORTED_MODULE_22__ = __webpack_require__(/*! ./transforms/grid */ \"./wp-content/plugins/vf-gutenberg/blocks/vf-core/transforms/grid.js\");\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\nfunction _toConsumableArray(arr) { return _arrayWithoutHoles(arr) || _iterableToArray(arr) || _nonIterableSpread(); }\n\nfunction _nonIterableSpread() { throw new TypeError(\"Invalid attempt to spread non-iterable instance\"); }\n\nfunction _iterableToArray(iter) { if (Symbol.iterator in Object(iter) || Object.prototype.toString.call(iter) === \"[object Arguments]\") return Array.from(iter); }\n\nfunction _arrayWithoutHoles(arr) { if (Array.isArray(arr)) { for (var i = 0, arr2 = new Array(arr.length); i < arr.length; i++) { arr2[i] = arr[i]; } return arr2; } }\n\nfunction ownKeys(object, enumerableOnly) { var keys = Object.keys(object); if (Object.getOwnPropertySymbols) { var symbols = Object.getOwnPropertySymbols(object); if (enumerableOnly) symbols = symbols.filter(function (sym) { return Object.getOwnPropertyDescriptor(object, sym).enumerable; }); keys.push.apply(keys, symbols); } return keys; }\n\nfunction _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i] != null ? arguments[i] : {}; if (i % 2) { ownKeys(source, true).forEach(function (key) { _defineProperty(target, key, source[key]); }); } else if (Object.getOwnPropertyDescriptors) { Object.defineProperties(target, Object.getOwnPropertyDescriptors(source)); } else { ownKeys(source).forEach(function (key) { Object.defineProperty(target, key, Object.getOwnPropertyDescriptor(source, key)); }); } } return target; }\n\nfunction _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }\n\n/**\nBlock Name: Grid\n*/\n\n\n\n\n\n\n\n\nvar defaults = Object(_hooks_use_vf_defaults__WEBPACK_IMPORTED_MODULE_20__[\"default\"])();\nvar ver = '1.0.0';\nvar MIN_COLUMNS = 1;\nvar MAX_COLUMNS = 6;\n\nvar settings = _objectSpread({}, defaults, {\n  name: 'vf/grid',\n  title: Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_19__[\"__\"])('VF Grid'),\n  category: 'vf/core',\n  description: Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_19__[\"__\"])('Visual Framework (core)'),\n  attributes: _objectSpread({}, defaults.attributes, {\n    placeholder: {\n      type: 'integer',\n      default: 0\n    },\n    columns: {\n      type: 'integer',\n      default: 0\n    }\n  })\n});\n\nsettings.save = function (props) {\n  if (props.attributes.placeholder === 1) {\n    return null;\n  }\n\n  var columns = props.attributes.columns;\n  var className = \"vf-grid | vf-grid__col-\".concat(columns);\n  return wp.element.createElement(\"div\", {\n    className: className\n  }, wp.element.createElement(_wordpress_block_editor__WEBPACK_IMPORTED_MODULE_16__[\"InnerBlocks\"].Content, null));\n};\n\nvar withGridDispatch = function withGridDispatch(Edit) {\n  return Object(_wordpress_data__WEBPACK_IMPORTED_MODULE_18__[\"withDispatch\"])(function (dispatch, ownProps, _ref) {\n    var select = _ref.select;\n\n    var _select = select('core/block-editor'),\n        getBlocks = _select.getBlocks;\n\n    var _dispatch = dispatch('core/block-editor'),\n        replaceInnerBlocks = _dispatch.replaceInnerBlocks; // `columns` attribute `onChange` callback\n\n\n    var setColumns = function setColumns(newColumns) {\n      var prevColumns = ownProps.attributes.columns; // Merge inner blocks when number of columns is reduced\n\n      if (newColumns < prevColumns) {\n        var columnBlocks = getBlocks(ownProps.clientId);\n        var mergeBlocks = [];\n\n        for (var i = newColumns - 1; i < prevColumns; i++) {\n          mergeBlocks.push.apply(mergeBlocks, _toConsumableArray(columnBlocks[i].innerBlocks));\n        }\n\n        replaceInnerBlocks(columnBlocks[newColumns - 1].clientId, mergeBlocks, false);\n      } // Update block attributes\n\n\n      ownProps.setAttributes({\n        columns: newColumns,\n        placeholder: 0\n      });\n    };\n\n    return {\n      setColumns: setColumns\n    };\n  })(Edit);\n};\n\nsettings.edit = withGridDispatch(function (props) {\n  var _props$attributes = props.attributes,\n      columns = _props$attributes.columns,\n      placeholder = _props$attributes.placeholder; // Ensure version is encoded in post content\n\n  props.setAttributes({\n    ver: ver\n  }); // Turn on setup placeholder if no columns are defined\n\n  if (columns === 0) {\n    props.setAttributes({\n      placeholder: 1\n    });\n  } // Setup placeholder fields\n\n\n  var fields = [{\n    control: 'columns',\n    min: MIN_COLUMNS,\n    max: MAX_COLUMNS,\n    value: columns,\n    onChange: props.setColumns\n  }]; // Return setup placeholder\n\n  if (placeholder === 1) {\n    return wp.element.createElement(\"div\", {\n      className: \"vf-block vf-block--placeholder \".concat(props.className)\n    }, wp.element.createElement(_wordpress_components__WEBPACK_IMPORTED_MODULE_17__[\"Placeholder\"], {\n      label: Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_19__[\"__\"])('VF Grid'),\n      icon: 'admin-generic'\n    }, wp.element.createElement(_vf_block_block_fields__WEBPACK_IMPORTED_MODULE_21__[\"default\"], {\n      fields: fields\n    })));\n  } // Amend fields for inspector\n\n\n  fields[0].help = Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_19__[\"__\"])('Content may be reorganised when columns are reduced.'); // Return inner blocks and inspector controls\n\n  return wp.element.createElement(react__WEBPACK_IMPORTED_MODULE_15__[\"Fragment\"], null, wp.element.createElement(_wordpress_block_editor__WEBPACK_IMPORTED_MODULE_16__[\"InspectorControls\"], null, wp.element.createElement(_wordpress_components__WEBPACK_IMPORTED_MODULE_17__[\"PanelBody\"], {\n    title: Object(_wordpress_i18n__WEBPACK_IMPORTED_MODULE_19__[\"__\"])('Settings'),\n    initialOpen: true\n  }, wp.element.createElement(_vf_block_block_fields__WEBPACK_IMPORTED_MODULE_21__[\"default\"], {\n    fields: fields\n  }))), wp.element.createElement(\"div\", {\n    className: 'vf-block-grid',\n    \"data-ver\": ver,\n    \"data-columns\": columns\n  }, wp.element.createElement(_wordpress_block_editor__WEBPACK_IMPORTED_MODULE_16__[\"InnerBlocks\"], {\n    allowedBlocks: ['vf/grid-column'],\n    template: Array(columns).fill(['vf/grid-column']),\n    templateLock: \"all\"\n  })));\n}); // Block transforms\n\nsettings.transforms = {\n  from: [Object(_transforms_grid__WEBPACK_IMPORTED_MODULE_22__[\"fromColumns\"])('core/columns', 'vf/grid', MIN_COLUMNS, MAX_COLUMNS), Object(_transforms_grid__WEBPACK_IMPORTED_MODULE_22__[\"fromColumns\"])('vf/embl-grid', 'vf/grid', MIN_COLUMNS, MAX_COLUMNS)]\n};\n/* harmony default export */ __webpack_exports__[\"default\"] = (settings);\n\n//# sourceURL=webpack:///./wp-content/plugins/vf-gutenberg/blocks/vf-core/vf-grid.jsx?");
 
 /***/ }),
